Renamed stage module to session
